+++ test/fomac/test_fomac.cpp
@@ -16,6 +16,7 @@
 #include <gtest/gtest.h>
 #include <new>
 #include <numbers>
+#include <optional>
 #include <qdmi/client.h>
 #include <ranges>
 #include <stdexcept>